Renamed: component ShowPaginatedProducts.vue
