+++ apps/frontend/src/components/page-products/ShowPaginatedProducts.vue
